:label: [rename] home 패키지 생성 및 Home Composable 위치를 home 패키지로 이동
