+++ app/src/main/java/com/example/clothesbyweather/MainActivity.kt
@@ -11,7 +11,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.example.clothesbyweather.ui.theme.ClothesByWeatherTheme
-import com.example.clothesbyweather.ui.theme.Home
+import com.example.clothesbyweather.home.Home
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
