+++ app/src/main/java/com/example/clothesbyweather/home/Home.kt
@@ -1,4 +1,4 @@
-package com.example.clothesbyweather.ui.theme
+package com.example.clothesbyweather.home
 
 import android.content.res.Configuration.UI_MODE_NIGHT_NO
 import androidx.compose.foundation.background
@@ -7,7 +7,6 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.ElevatedButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
@@ -17,6 +16,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.example.clothesbyweather.ui.theme.ClothesByWeatherTheme
 
 @Composable
 fun Home(modifier: Modifier) {
